Datamuse-API-similarwords
Uses the datamuse API to fetch similar words to the one provided in the input field. After getting the words from the API they are displayed in an li element contained in an ul. Clicking on one of the list items will prompt another API call pulling similar words to it.